Add branch prefix whitelist to webhook
R10K supports the ability to provision environments from a source
without a prefix. However, the webhook currently supports provisioning
prefixed sources only. In some cases, this is less than ideal.

This commit adds an optional prefix whitelist for the webhook. This
whitelist will accept a comma separated list of sources. When the prefix
command returns a prefix on the whitelist, R10K will be invoked without
the prefixed environment. This option is only effective when webhook
prefix awareness is enabled.